requirements.txt
streamlit
pandas
numpy
snowflake-connector-python
snowflake-snowpark-python[pandas]
boto3
botocore
